Fix comb stage by putting delay _before_ sum, not after.
Previously the comb equation was:
y[n] =  x[n] - y[n-1]
This was _wrong_, as this put the delay _after_ the sum.
These changes should make the equation:
y[n] = x[n] - x[n-1]

To do this, we needed to expand the combs array to include the _input_ to the comb section, and rejiggle the delay copy to suit.
We also needed to include the first stage in the for loop as it was no longer processed externaly.
The expansion of the combs array changes the meaning of combs[0] from:
Before: combs[R][0] = output of first comb
After: combs[R][0] = _input_ of first comb
Loops have been changed to suit, and the frequency response now looks good.